% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{package}
\name{shinyPostgreSQL}
\alias{shinyPostgreSQL}
\title{shinyPostgreSQL: Access PostgreSQL databases with R Shiny}
\description{
A shiny module to authenticate your R Shiny Application with a PostgreSQL database.
}
